refactor: make block inspector a sub component
Playing around with how to compartmentalize state etc
