+++ src/block_inspector/mod.rs
@@ -1,0 +1,7 @@
+mod state;
+mod update;
+mod view;
+
+pub use state::*;
+pub use update::*;
+pub use view::*;
